+++ pipeline/checksums/checksums_test.go
@@ -11,7 +11,7 @@ import (
 )
 
 func TestDescription(t *testing.T) {
-	assert.NotEmpty(t, Pipe{}.Description())
+	assert.NotEmpty(t, Pipe{}.String())
 }
 
 func TestPipe(t *testing.T) {
