+++ pipeline/cleandist/dist.go
@@ -15,8 +15,8 @@ import (
 type Pipe struct{}
 
 // Description of the pipe
-func (Pipe) Description() string {
-	return "Checking ./dist"
+func (Pipe) String() string {
+	return "checking ./dist"
 }
 
 // Run the pipe
