+++ pipeline/cleandist/dist_test.go
@@ -59,5 +59,5 @@ func TestEmptyDistExists(t *testing.T) {
 }
 
 func TestDescription(t *testing.T) {
-	assert.NotEmpty(t, Pipe{}.Description())
+	assert.NotEmpty(t, Pipe{}.String())
 }
