+++ pipeline/defaults/defaults_test.go
@@ -10,7 +10,7 @@ import (
 )
 
 func TestDescription(t *testing.T) {
-	assert.NotEmpty(t, Pipe{}.Description())
+	assert.NotEmpty(t, Pipe{}.String())
 }
 
 func TestFillBasicData(t *testing.T) {
