+++ pipeline/fpm/fpm_test.go
@@ -14,7 +14,7 @@ import (
 )
 
 func TestDescription(t *testing.T) {
-	assert.NotEmpty(t, Pipe{}.Description())
+	assert.NotEmpty(t, Pipe{}.String())
 }
 
 func TestRunPipeNoFormats(t *testing.T) {
