+++ pipeline/git/git.go
@@ -20,8 +20,8 @@ import (
 type Pipe struct{}
 
 // Description of the pipe
-func (Pipe) Description() string {
-	return "Getting and validating git state"
+func (Pipe) String() string {
+	return "getting and validating git state"
 }
 
 // Run the pipe
