+++ pipeline/release/release_test.go
@@ -15,7 +15,7 @@ import (
 )
 
 func TestPipeDescription(t *testing.T) {
-	assert.NotEmpty(t, Pipe{}.Description())
+	assert.NotEmpty(t, Pipe{}.String())
 }
 
 func TestRunPipe(t *testing.T) {
