+++ pipeline/release/remote.go
@@ -1,4 +1,4 @@
-package defaults
+package release
 
 import (
 	"strings"
