+++ pipeline/snapcraft/snapcraft_test.go
@@ -15,7 +15,7 @@ import (
 )
 
 func TestDescription(t *testing.T) {
-	assert.NotEmpty(t, Pipe{}.Description())
+	assert.NotEmpty(t, Pipe{}.String())
 }
 
 func TestRunPipeMissingInfo(t *testing.T) {
