+++ pipeline/snapshot/package.go
@@ -1,0 +1,2 @@
+// Package snapshot provides the snapshoting functionaly to goreleaser
+package snapshot
