+++ pipeline/snapshot/snapshot.go
@@ -1,0 +1,20 @@
+package snapshot
+
+import (
+	"github.com/goreleaser/goreleaser/context"
+)
+
+// Pipe for checksums
+type Pipe struct{}
+
+func (Pipe) String() string {
+	return "generating changelog"
+}
+
+// Default sets the pipe defaults
+func (Pipe) Default(ctx *context.Context) error {
+	if ctx.Config.Snapshot.NameTemplate == "" {
+		ctx.Config.Snapshot.NameTemplate = "SNAPSHOT-{{ .Commit }}"
+	}
+	return nil
+}
